backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
boto3==1.34.21
openai==1.51.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
